requirements.txt
streamlit
python-dotenv
pypdf
groq
